+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- DisableCursorListener component
 - Mouse IsInsideGameView extension
 
 ## [1.2.0] - 2025-08-17